Agrego arreglos en reportes

+++ reportes/src/main/java/com/reportes/reportes/DTOs/ReporteFacturacion.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -14,5 +15,5 @@ import java.time.LocalDateTime;
 public class ReporteFacturacion {
     private LocalDateTime startDate;
     private LocalDateTime endDate;
-    private double cantFacturado;
+    private BigDecimal cantFacturado;
 }
